requirements.txt
django~=4.0.5
requests~=2.28.0
djangorestframework~=3.13.1
djangorestframework-jwt~=1.11.0